Switch to the hook method
