+++ src/utils/fetchers/useFetchZenikaPageData.tsx
@@ -1,5 +1,4 @@
-import { gql, useQuery } from "@apollo/client";
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
+import { useQuery } from "@apollo/client";
 import {
   GetSkillsAndDesiresByAgencyQuery,
   GetSkillsAndDesiresQuery,
@@ -9,21 +8,21 @@ import {
   GET_SKILLS_AND_DESIRES_BY_AGENCY_QUERY,
 } from "../../graphql/queries/skills";
 
-const fetchZenikaPageDataByAgency = (
-  email: string,
-  agency: string
-) => {
-  const { data: skillsData, error } = useQuery<
-    GetSkillsAndDesiresByAgencyQuery
-  >(GET_SKILLS_AND_DESIRES_BY_AGENCY_QUERY, {
-    variables: { email, agency },
-  });
+const fetchZenikaPageDataByAgency = (email: string, agency: string) => {
+  const { data: skillsData, error } =
+    useQuery<GetSkillsAndDesiresByAgencyQuery>(
+      GET_SKILLS_AND_DESIRES_BY_AGENCY_QUERY,
+      {
+        variables: { email, agency },
+      }
+    );
   const homePanelData = skillsData?.Category?.map((data) => ({
     x: data?.x,
     y: data?.y,
     color: data?.color,
     name: data?.label,
-    count: data?.AgenciesAverageCurrentSkillsAndDesires_aggregate.aggregate.count,
+    count:
+      data?.AgenciesAverageCurrentSkillsAndDesires_aggregate.aggregate.count,
     context: "zenika",
     data: data?.AgenciesAverageCurrentSkillsAndDesires.map((skill, i) => ({
       x: skill.averageSkillLevel,
@@ -54,7 +53,8 @@ const fetchZenikaPageData = (email: string) => {
     y: data?.y,
     color: data?.color,
     name: data?.label,
-    count: data?.ZenikasAverageCurrentSkillsAndDesires_aggregate.aggregate.count,
+    count:
+      data?.ZenikasAverageCurrentSkillsAndDesires_aggregate.aggregate.count,
     context: "zenika",
     data: data?.ZenikasAverageCurrentSkillsAndDesires.map((skill, i) => ({
       x: skill.averageSkillLevel,
@@ -73,7 +73,7 @@ const fetchZenikaPageData = (email: string) => {
 
 export const useFetchZenikaPageData = (
   email: string,
-  agency: string | undefined,
+  agency: string | undefined
 ) =>
   agency && agency !== "World"
     ? fetchZenikaPageDataByAgency(email, agency)
